backend/requirements.txt
pip>=25.0.1
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
isodate==0.7.2
paramiko==3.5.1
pillow==10.4.0
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
typing_extensions==4.12.2
urllib3==2.2.3
